requirements.txt
flask
PyPDF2
python-docx
transformers
torch
concurrent-log-handler
